+++ chain-signatures/chain-integration-core/src/lib.rs
@@ -1,12 +1,13 @@
 //! This crate provides the core traits and types for implementing chain indexers and streams for different blockchains.
 
 mod indexer;
+mod publish;
 mod state;
 mod telemetry;
-mod publish;
 pub mod utils;
 
 pub use indexer::{ChainIndexer, ChainStream};
+pub use publish::{ChainPublisher, PublishAction};
 pub use state::StateManager;
 pub use telemetry::{ChainTelemetry, NoopChainTelemetry};
