+++ chain-signatures/node/src/cli/mod.rs
@@ -17,7 +17,7 @@ use crate::protocol::state::Node;
 use crate::protocol::sync::SyncTask;
 use crate::protocol::Chain;
 use crate::protocol::{spawn_system_metrics, MpcSignProtocol};
-use crate::rpc::{self, ChainPublisher, ContractStateWatcher, NearClient, RpcExecutor};
+use crate::rpc::{self, ContractStateWatcher, NearClient, RpcExecutor};
 use crate::storage::checkpoint_storage::CheckpointStorage;
 use crate::storage::triple_storage::TriplePair;
 use crate::stream::run_stream;
@@ -29,6 +29,7 @@ use deadpool_redis::Runtime;
 use enum_map::EnumMap;
 use k256::sha2::Sha256;
 use local_ip_address::local_ip;
+use mpc_chain_integration_core::ChainPublisher;
 use mpc_keys::hpke;
 use mpc_primitives::CheckpointDigest;
 use near_account_id::AccountId;
