+++ chain-signatures/node/src/indexer_sol/client.rs
@@ -26,8 +26,10 @@ use std::str::FromStr;
 use std::sync::Arc;
 use std::time::Duration;
 
-use crate::rpc::{ChainPublisher, PublishAction};
-use mpc_chain_integration_core::utils::retry::{retry_rpc, RetryConfig};
+use mpc_chain_integration_core::{
+    utils::retry::{retry_rpc, RetryConfig},
+    ChainPublisher, PublishAction,
+};
 
 const MAX_SIGNATURES_FOR_FAST_CATCHUP: usize = 1000;
