+++ chain-signatures/node/src/rpc/ethereum.rs
@@ -1,4 +1,3 @@
-use super::{ChainPublisher, PublishAction};
 use crate::indexer_eth::abi::ChainSignatures;
 use crate::indexer_eth::EthConfig;
 use alloy::network::EthereumWallet;
@@ -12,7 +11,10 @@ use alloy::providers::{
 use alloy::rpc::types::TransactionReceipt;
 use alloy_signer_local::PrivateKeySigner;
 use k256::elliptic_curve::{point::AffineCoordinates, sec1::ToEncodedPoint};
-use mpc_chain_integration_core::utils::retry::{retry_rpc, RetryConfig};
+use mpc_chain_integration_core::{
+    utils::retry::{retry_rpc, RetryConfig},
+    ChainPublisher, PublishAction,
+};
 use mpc_primitives::{SignId, Signature};
 use std::collections::HashMap;
 use std::str::FromStr;
